bootc/requirements.txt
maskpass
passlib
PyYAML